code/requirements.txt
numpy
scipy
matplotlib
Pillow
scikit-image
scikit-fmm
pynrrd
nibabel